requirements.txt
google-analytics-data==0.18.7
google-api-python-client==2.108.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2
requests==2.31.0
slack-webhook==1.0.1